app/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Azure — identity and services
azure-identity==1.19.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.23.0

# Auth — Entra ID (MSAL)
msal==1.31.0

# Configuration
pydantic-settings==2.6.0
pydantic==2.9.0

# HTTP client (for MSAL token validation)
httpx==0.27.2

# Data processing (pipeline)
pandas==2.2.3
python-multipart==0.0.12

# Security
cryptography==43.0.3

# Statistics (Report 7 chi-square test)
scipy==1.14.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
